Fix fp16 GradScaler crash: auto-detect model dtype in trainers
- Change default model loading dtype from float16 to float32
- Add dtype detection in CPT and SFT trainers: if model is already
  fp16, disable the GradScaler (fp16=False) to avoid the
  "Attempting to unscale FP16 gradients" error
- Both paths work: fp32 model + fp16 mixed precision (standard),
  or fp16 model + no scaler (when user loads model in fp16)

+++ mathphd_plus_plus/models/base_model.py
@@ -41,7 +41,7 @@ def load_tokenizer(
 
 def load_model(
     model_name: str = "Qwen/Qwen2.5-0.5B",
-    torch_dtype: str = "float16",
+    torch_dtype: str = "float32",
     gradient_checkpointing: bool = True,
     device_map: str = "auto",
     quantization: Optional[str] = None,  # None, "4bit", "8bit"
